+++ lab_report.go
@@ -61,10 +61,12 @@ var (
 	labReportResultFieldMaxReferenceRange = big.NewInt(1 << 6)
 	labReportResultFieldMinReferenceRange = big.NewInt(1 << 7)
 	labReportResultFieldSourcePanelName   = big.NewInt(1 << 8)
-	labReportResultFieldLoincMatches      = big.NewInt(1 << 9)
-	labReportResultFieldInterpretation    = big.NewInt(1 << 10)
-	labReportResultFieldIsAboveMaxRange   = big.NewInt(1 << 11)
-	labReportResultFieldIsBelowMinRange   = big.NewInt(1 << 12)
+	labReportResultFieldIsSensitive       = big.NewInt(1 << 9)
+	labReportResultFieldLoincMatches      = big.NewInt(1 << 10)
+	labReportResultFieldLoincMatchStatus  = big.NewInt(1 << 11)
+	labReportResultFieldInterpretation    = big.NewInt(1 << 12)
+	labReportResultFieldIsAboveMaxRange   = big.NewInt(1 << 13)
+	labReportResultFieldIsBelowMinRange   = big.NewInt(1 << 14)
 )
 
 type LabReportResult struct {
@@ -80,7 +82,11 @@ type LabReportResult struct {
 	MaxReferenceRange *float64             `json:"max_reference_range,omitempty" url:"max_reference_range,omitempty"`
 	MinReferenceRange *float64             `json:"min_reference_range,omitempty" url:"min_reference_range,omitempty"`
 	SourcePanelName   *string              `json:"source_panel_name,omitempty" url:"source_panel_name,omitempty"`
-	LoincMatches      []*LoincMatch        `json:"loinc_matches,omitempty" url:"loinc_matches,omitempty"`
+	// ℹ️ This enum is non-exhaustive.
+	IsSensitive  *LabReportResultIsSensitive `json:"is_sensitive,omitempty" url:"is_sensitive,omitempty"`
+	LoincMatches []*LoincMatch               `json:"loinc_matches,omitempty" url:"loinc_matches,omitempty"`
+	// ℹ️ This enum is non-exhaustive.
+	LoincMatchStatus *LabReportResultLoincMatchStatus `json:"loinc_match_status,omitempty" url:"loinc_match_status,omitempty"`
 	// ℹ️ This enum is non-exhaustive.
 	Interpretation  *Interpretation `json:"interpretation,omitempty" url:"interpretation,omitempty"`
 	IsAboveMaxRange *bool           `json:"is_above_max_range,omitempty" url:"is_above_max_range,omitempty"`
@@ -156,11 +162,25 @@ func (l *LabReportResult) GetSourcePanelName() *string {
 	return l.SourcePanelName
 }
 
+func (l *LabReportResult) GetIsSensitive() *LabReportResultIsSensitive {
+	if l == nil {
+		return nil
+	}
+	return l.IsSensitive
+}
+
 func (l *LabReportResult) GetLoincMatches() []*LoincMatch {
 	if l == nil {
 		return nil
 	}
 	return l.LoincMatches
+}
+
+func (l *LabReportResult) GetLoincMatchStatus() *LabReportResultLoincMatchStatus {
+	if l == nil {
+		return nil
+	}
+	return l.LoincMatchStatus
 }
 
 func (l *LabReportResult) GetInterpretation() *Interpretation {
@@ -261,11 +281,25 @@ func (l *LabReportResult) SetSourcePanelName(sourcePanelName *string) {
 	l.require(labReportResultFieldSourcePanelName)
 }
 
+// SetIsSensitive sets the IsSensitive field and marks it as non-optional;
+// this prevents an empty or null value for this field from being omitted during serialization.
+func (l *LabReportResult) SetIsSensitive(isSensitive *LabReportResultIsSensitive) {
+	l.IsSensitive = isSensitive
+	l.require(labReportResultFieldIsSensitive)
+}
+
 // SetLoincMatches sets the LoincMatches field and marks it as non-optional;
 // this prevents an empty or null value for this field from being omitted during serialization.
 func (l *LabReportResult) SetLoincMatches(loincMatches []*LoincMatch) {
 	l.LoincMatches = loincMatches
 	l.require(labReportResultFieldLoincMatches)
+}
+
+// SetLoincMatchStatus sets the LoincMatchStatus field and marks it as non-optional;
+// this prevents an empty or null value for this field from being omitted during serialization.
+func (l *LabReportResult) SetLoincMatchStatus(loincMatchStatus *LabReportResultLoincMatchStatus) {
+	l.LoincMatchStatus = loincMatchStatus
+	l.require(labReportResultFieldLoincMatchStatus)
 }
 
 // SetInterpretation sets the Interpretation field and marks it as non-optional;
@@ -329,6 +363,57 @@ func (l *LabReportResult) String() string {
 		return value
 	}
 	return fmt.Sprintf("%#v", l)
+}
+
+// ℹ️ This enum is non-exhaustive.
+type LabReportResultIsSensitive string
+
+const (
+	LabReportResultIsSensitiveSensitive   LabReportResultIsSensitive = "sensitive"
+	LabReportResultIsSensitiveInsensitive LabReportResultIsSensitive = "insensitive"
+	LabReportResultIsSensitiveUnknown     LabReportResultIsSensitive = "unknown"
+)
+
+func NewLabReportResultIsSensitiveFromString(s string) (LabReportResultIsSensitive, error) {
+	switch s {
+	case "sensitive":
+		return LabReportResultIsSensitiveSensitive, nil
+	case "insensitive":
+		return LabReportResultIsSensitiveInsensitive, nil
+	case "unknown":
+		return LabReportResultIsSensitiveUnknown, nil
+	}
+	var t LabReportResultIsSensitive
+	return "", fmt.Errorf("%s is not a valid %T", s, t)
+}
+
+func (l LabReportResultIsSensitive) Ptr() *LabReportResultIsSensitive {
+	return &l
+}
+
+type LabReportResultLoincMatchStatus string
+
+const (
+	LabReportResultLoincMatchStatusAutoMatch   LabReportResultLoincMatchStatus = "auto_match"
+	LabReportResultLoincMatchStatusNeedsReview LabReportResultLoincMatchStatus = "needs_review"
+	LabReportResultLoincMatchStatusNoMatch     LabReportResultLoincMatchStatus = "no_match"
+)
+
+func NewLabReportResultLoincMatchStatusFromString(s string) (LabReportResultLoincMatchStatus, error) {
+	switch s {
+	case "auto_match":
+		return LabReportResultLoincMatchStatusAutoMatch, nil
+	case "needs_review":
+		return LabReportResultLoincMatchStatusNeedsReview, nil
+	case "no_match":
+		return LabReportResultLoincMatchStatusNoMatch, nil
+	}
+	var t LabReportResultLoincMatchStatus
+	return "", fmt.Errorf("%s is not a valid %T", s, t)
+}
+
+func (l LabReportResultLoincMatchStatus) Ptr() *LabReportResultLoincMatchStatus {
+	return &l
 }
 
 // ℹ️ This enum is non-exhaustive.